adjusted classes for dataset and ohcl dataframes to make more sense
