+++ main/prod/python/dataset.py
@@ -1,23 +1,29 @@
-import pandas as pd
-from prod.python.binance import Binance as binance
-from common.python import database_operations as db
+import pandas_ta as ta
 
 class Dataset():
-    def __init__(self, pair, interval, startTime, endTime):
-        self.pair = pair
-        self.interval = interval
-        self.startTime = startTime
-        self.endTime = endTime
-    
-    def populate_ohlc(self):
-        self.dataset = binance().get_extended_kline(self.pair, self.interval, self.startTime, self.endTime)
+    def __init__(self):
+        #initializing strategy (group of indicators)
+        #a strategy for pandas_ta is a group of indicators that will be added to the dataset
+        self.strategy = ta.Strategy(
+            name="grouping_indicators",
+            ta=[{}]
+        )
 
-    def get_new_row(self):
-        return binance().get_kline(self.pair, self.interval, None, None, 2).iloc[0]
+    def calc_indicator(self, df, indicator: str, **kwargs):
+        try:
+            return df.ta(kind=indicator, append=True, **kwargs)
+        except Exception as e:
+            raise RuntimeError(f'Indicator "{indicator}" error with exception: {e}')
 
-    def append_row_to_df(self, row):
-        self.dataset.loc[len(self.dataset)] = row
+    def add_indicator_to_strategy(self, indicator):
+        #indicator has to be a dictionary like: {"kind": "rsi", "length": 22}
+        self.strategy.ta.append(indicator)
 
-    def update_dataset(self, row):
-        self.append_row_to_df(row)
-        self.dataset.drop([0])
+    def apply_strategy_to_df(self, df):
+        df.ta.strategy(self.strategy)
+
+    def join_indicator_to_dataset(self, df, indicator):
+        return df.join(indicator)
+
+
+
